This update closes the following issues:

- closes #7 - replace mackup
- closes #14 - replace open apps before setup
- closes #14 - add --dry-run option
- closes #18 - brew cask small apps

Plus it improves on a lot of things:

- better output
- try now handles sudo
- new args
- better argument support
- faster CI
- rsync a lot of config from Library/Preferences
- etc. 🎉

+++ scripts/brew.sh
@@ -3,17 +3,15 @@
 # Continue on error
 set +e
 
-# Install command-line tools using Homebrew.
+if [[ ! -n $CI_TEST ]]; then
+	echo ""
+	echo "› Update brew"
+	try brew update
+	try brew upgrade
+fi
 
-# Make sure we’re using the latest Homebrew.
-try brew update
-
-# Upgrade any already-installed formulae.
-try brew upgrade
-
-# List already available formulae & casks
+# List already available formulae
 available_formulae=$(brew list)
-available_casks=$(brew cask list)
 
 typeset -U formulae
 formulae=(
@@ -21,7 +19,7 @@ formulae=(
 	# Don’t forget to add `$(brew --prefix coreutils)/libexec/gnubin` to `$PATH`
 	coreutils
 	# Install some other useful utilities like `sponge`
-	moreutils
+	# moreutils
 	# Install GNU `find`, `locate`, `updatedb`, and `xargs`, `g`-prefixed
 	findutils
 	# Install macOS utils
@@ -84,60 +82,19 @@ formulae=(
 	swiftlint
 	clang-format
 	cppcheck
-
-	# Install osx-cross formulae
-	osx-cross/arm/arm-gcc-bin
-	osx-cross/avr/avr-gcc
 )
 
-# Install formulae
+echo ""
+echo "› Install formulae"
 for formula in $formulae ; do
 	if [[ ! $available_formulae =~ $formula ]]; then
 		try brew install $formula
 	fi
 done
 
-typeset -U casks
-casks=(
-	1password
-	alfred
-	appcleaner
-	arduino
-	brave-browser
-	coolterm
-	dropbox
-	fantastical
-	# font-source-code-pro
-	google-chrome
-	gpg-suite-no-mail
-	iterm2
-	macdown
-	# mactex-no-gui
-	qlcolorcode
-	qlimagesize
-	qlmarkdown
-	qlprettypatch
-	qlstephen
-	quicklook-csv
-	quicklook-json
-	rectangle
-	slack
-	spotify
-	sublime-text
-	transmission
-	# the-unarchiver
-	visual-studio-code
-	vlc
-	whatsapp
-)
-
-# Install casks
-for cask in $casks; do
-	if [[ ! $available_casks =~ $cask ]]; then
-		try brew cask install $cask
-	fi
-done
-
-# Remove outdated versions from the cellar
-try brew cleanup -s
-try rm -rf "$(brew --cache)"
+if [[ ! -n $CI_TEST ]]; then
+	echo ""
+	echo "› Cleanup brew & remove cache"
+	try brew cleanup -s
+	try rm -rf "$(brew --cache)"
+fi
